requirements.txt
aioredis==1.3.1
asgiref==3.2.7
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.6
django-crispy==1.8.1
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
djangorestframework==3.11.0
Flask==1.1.2
Flask-MySQL==1.5.1
hiredis==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
netifaces==0.10.4
olefile==0.45.1
pexpect==4.2.1
pgi==0.0.11.2
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyotp==2.3.0
pyTelegramBotAPI==3.7.1
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
systemd-python==234
twilio==6.39.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
